requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
django-haystack==3.1.1
djangorestframework==3.12.4
elasticsearch==1.9.0
pyelasticsearch==1.4.1
pytz==2021.3
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
